move dynamic and gvk fully to kube-core
