+++ kube-core/src/dynamic.rs
