+++ kube/src/api/mod.rs
@@ -9,21 +9,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Deserialize, Serialize, Default, Debug)]
 pub struct NotUsed {}
 
-pub(crate) use kube_core::params;
-use kube_core::request;
-pub use params::{
-    DeleteParams, ListParams, Patch, PatchParams, PostParams, Preconditions, PropagationPolicy,
-};
-pub use request::Request;
-
 pub(crate) mod typed;
 pub use typed::Api;
-
-mod gvk;
-pub use gvk::{GroupVersionKind, GroupVersionResource};
-
-mod dynamic;
-pub use dynamic::{ApiResource, DynamicObject};
 
 #[cfg(feature = "ws")] mod remote_command;
 #[cfg(feature = "ws")] pub use remote_command::AttachedProcess;
@@ -36,7 +23,15 @@ pub use subresource::{EvictParams, Evictable, LogParams, Loggable, ScaleSpec, Sc
 pub(crate) mod object;
 pub use self::object::{Object, ObjectList, WatchEvent};
 
-use kube_core::metadata;
-pub use self::metadata::{ListMeta, ObjectMeta, Resource, ResourceExt, TypeMeta};
+// Re-exports from kube-core
+pub use kube_core::gvk::{GroupVersionKind, GroupVersionResource};
+pub use kube_core::dynamic::{ApiResource, DynamicObject};
+pub use kube_core::metadata::{ListMeta, ObjectMeta, Resource, ResourceExt, TypeMeta};
+pub use kube_core::request::Request;
+pub(crate) use kube_core::params;
+pub use params::{
+    DeleteParams, ListParams, Patch, PatchParams, PostParams, Preconditions, PropagationPolicy,
+};
+
 
 #[cfg(feature = "admission")] pub mod admission;
